chore: lint, fixes, etc
