+++ js/packages/api/index.gen.js
@@ -155,6 +155,11 @@ export const berty = {
 		},
 		ShareableContact: jsonPb.lookup('.berty.types.ShareableContact'),
 	},
+	chat: {
+		ChatService: jsonPb.lookup('.berty.chat.ChatService'),
+		InstanceShareableBertyID: jsonPb.lookup('.berty.chat.InstanceShareableBertyID'),
+		DevShareInstanceBertyID: jsonPb.lookup('.berty.chat.DevShareInstanceBertyID'),
+	},
 }
 export const gogoproto = {
 	goprotoEnumPrefix: jsonPb.lookup('.gogoproto.goprotoEnumPrefix'),
